### 2023-10-02 ae64206
Bump dscribe from 2.0.0 to 2.1.0
Bumps [dscribe](https://singroup.github.io/dscribe/) from 2.0.0 to 2.1.0.

---
updated-dependencies:
- dependency-name: dscribe
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
