### 2023-10-02 #5915 @jgostick
Please take it...I always try to deflect DFT / MD things.  There was something on the last email about image analysis that looked more relevant to me, so I'll go add my name to that.
