### 2023-10-02 #5915 @rkurchin
👋 @ireaml and @prashungorai (or possibly one of your students?), would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
