+++ .pre-commit-config.yaml
@@ -9,15 +9,15 @@ repos:
   - id: end-of-file-fixer
   - id: trailing-whitespace
 - repo: https://github.com/PyCQA/autoflake
-  rev: v2.2.0
+  rev: v2.2.1
   hooks:
   - id: autoflake
 - repo: https://github.com/psf/black
-  rev: 23.3.0
+  rev: 23.9.1
   hooks:
   - id: black
 - repo: https://github.com/asottile/blacken-docs
-  rev: 1.15.0
+  rev: 1.16.0
   hooks:
   - id: blacken-docs
     additional_dependencies: [black]
@@ -27,13 +27,13 @@ repos:
   hooks:
   - id: isort
 - repo: https://github.com/pycqa/flake8
-  rev: 6.0.0
+  rev: 6.1.0
   hooks:
   - id: flake8
     entry: pflake8
     files: ^src/
     additional_dependencies:
-    - pyproject-flake8==6.0.0
+    - pyproject-flake8==6.1.0
     - flake8-bugbear==22.12.6
     - flake8-typing-imports==1.14.0
     - flake8-docstrings==1.6.0
@@ -52,7 +52,7 @@ repos:
   - id: rst-directive-colons
   - id: rst-inline-touching-normal
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v1.4.1
+  rev: v1.5.1
   hooks:
   - id: mypy
     files: ^pymatgen/
@@ -63,7 +63,7 @@ repos:
     - tokenize-rt==4.1.0
     - types-pkg_resources==0.1.2
 - repo: https://github.com/codespell-project/codespell
-  rev: v2.2.5
+  rev: v2.2.6
   hooks:
   - id: codespell
     name: codespell
@@ -72,8 +72,8 @@ repos:
     language: python
     types: [text]
     args: [
-      --ignore-words-list, 'titel,statics,ba,nd,te,mater,commun,vise',
-      --skip, "*.ipynb,./tests"
+      --ignore-words-list, 'titel,statics,ba,nd,te,mater,commun,vise,dscribe',
+      --skip, "*.ipynb,./tests,*paper*",
       ]
 - repo: https://github.com/kynan/nbstripout
   rev: 0.6.1
