+++ tests/test_thermo.py
@@ -23,6 +23,56 @@ from pymatgen.analysis.defects.thermo import (
 )
 
 
+@pytest.fixture(scope="module")
+def formation_energy_diagram(
+    data_Mg_Ga, defect_entries_and_plot_data_Mg_Ga, stable_entries_Mg_Ga_N
+):
+    bulk_vasprun = data_Mg_Ga["bulk_sc"]["vasprun"]
+    bulk_bs = bulk_vasprun.get_band_structure()
+    vbm = bulk_bs.get_vbm()["energy"]
+    bulk_entry = bulk_vasprun.get_computed_entry(inc_structure=False)
+    defect_entries, _ = defect_entries_and_plot_data_Mg_Ga
+
+    def_ent_list = list(defect_entries.values())
+    # test the constructor with materials project phase diagram
+    atomic_entries = list(
+        filter(lambda x: len(x.composition.elements) == 1, stable_entries_Mg_Ga_N)
+    )
+    pd = PhaseDiagram(stable_entries_Mg_Ga_N)
+
+    # test basic constructor
+    fed_ = FormationEnergyDiagram(
+        bulk_entry=bulk_entry,
+        defect_entries=def_ent_list,
+        vbm=vbm,
+        pd_entries=stable_entries_Mg_Ga_N,
+        inc_inf_values=True,  # include the two additional points at infinity
+    )
+    assert len(fed_.chempot_limits) == 5
+
+    # test the constructor with atomic entries
+    # this is the one we will use for the rest of the tests
+    fed = FormationEnergyDiagram.with_atomic_entries(
+        defect_entries=def_ent_list,
+        atomic_entries=atomic_entries,
+        vbm=vbm,
+        inc_inf_values=False,
+        phase_diagram=pd,
+        bulk_entry=bulk_entry,
+    )
+    assert len(fed.chempot_limits) == 3
+
+    # dataframe conversion
+    df = fed.as_dataframe()
+    assert df.shape == (4, 5)
+
+    # test that you can get the Ga-rich chempot
+    cp = fed.get_chempots(rich_element=Element("Ga"))
+    assert cp[Element("Ga")] == pytest.approx(0, abs=1e-2)
+
+    return fed
+
+
 def test_lower_envelope():
     # Test the lower envelope and transition code with a simple example
     lines = [[4, 12], [-1, 3], [-5, 4], [-2, 1], [3, 8], [-4, 14], [2, 12], [3, 8]]
@@ -68,58 +118,43 @@ def test_defect_entry(defect_entries_and_plot_data_Mg_Ga, data_Mg_Ga):
     assert def_entry.get_ediff() == pytest.approx(ediff, abs=1e-4)
 
 
-def test_formation_energy(
-    data_Mg_Ga, defect_entries_and_plot_data_Mg_Ga, stable_entries_Mg_Ga_N
-):
-    bulk_vasprun = data_Mg_Ga["bulk_sc"]["vasprun"]
-    bulk_bs = bulk_vasprun.get_band_structure()
-    vbm = bulk_bs.get_vbm()["energy"]
-    bulk_entry = bulk_vasprun.get_computed_entry(inc_structure=False)
-    defect_entries, plot_data = defect_entries_and_plot_data_Mg_Ga
-
-    def_ent_list = list(defect_entries.values())
-    fed = FormationEnergyDiagram(
-        bulk_entry=bulk_entry,
-        defect_entries=def_ent_list,
-        vbm=vbm,
-        pd_entries=stable_entries_Mg_Ga_N,
-        inc_inf_values=True,
-    )
-    assert len(fed.chempot_limits) == 5
-
-    # check that the constructor raises an error if `bulk_entry` is is provided for all defect entries
-    def_ents_w_bulk = copy.deepcopy(def_ent_list)
[…]
