### 2023-10-11 #5915 @rkurchin
@jmmshn, a question for you – would a reviewer be able to fully test the functionality of those code if they didn't themselves have access to a VASP license? To be clear, this is not at all a concern over eligibility for the journal, just one impacting reviewer selection.
