### 2023-10-11 #5915 @jmmshn
Hi @rkurchin, the overwhelming majority of the functionality should be testable without a VASP license.
Only the integrating with atomate2 requires VASP to fully test but that code currently lives in the atomate2 repos so it should not create any problems.
