### 2023-10-11 #5941 @editorialbot
[REVIEW]: pymatgen-analysis-defects: A Python package for analyzing point defects in crystalline materials
**Submitting author:** @jmmshn (<a href="http://orcid.org/0000-0002-2743-7531">Jimmy-Xuan Shen</a>)
**Repository:** https://github.com/materialsproject/pymatgen-analysis-defects
**Branch with paper.md** (empty if default branch):
**Version:** 2024.1.12
