### 2023-10-12 bb27313
Bump pymatgen from 2023.7.20 to 2023.10.11
Bumps [pymatgen](https://github.com/materialsproject/pymatgen) from 2023.7.20 to 2023.10.11.
- [Release notes](https://github.com/materialsproject/pymatgen/releases)
- [Changelog](https://github.com/materialsproject/pymatgen/blob/master/CHANGES.md)
- [Commits](https://github.com/materialsproject/pymatgen/compare/v2023.7.20...v2023.10.11)

---
updated-dependencies:
- dependency-name: pymatgen
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
