### 2023-10-12 d4b5d8b
Plot Competing Phases in 2D (#154)
* competing phases

* better ediff

* better ediff

* refactor tests

* docs

* docs

* docs

* dep

* dep
