+++ pyproject.toml
@@ -16,7 +16,7 @@ classifiers = [
   "Topic :: Scientific/Engineering",
 ]
 dependencies = [
-  "pymatgen>=2023.5.8",
+  "pymatgen>=v2023.10.11",
   "scikit-image>=0.19.3",
   "numpy"
 ]
@@ -37,10 +37,9 @@ docs = [
 pydefect = ["pydefect>=0.6.2"]
 
 strict = [
-  "pymatgen==2023.7.20",
+  "pymatgen==2023.10.11",
   "dscribe==2.1.0",
   "scikit-image==0.21.0",
-  "numpy==1.24.4",
 ]
 
 tests = ["pytest==7.4.2", "pytest-cov==4.1.0"]
