### 2023-10-16 21e31ae
Bump scikit-image from 0.21.0 to 0.22.0
Bumps [scikit-image](https://github.com/scikit-image/scikit-image) from 0.21.0 to 0.22.0.
- [Release notes](https://github.com/scikit-image/scikit-image/releases)
- [Changelog](https://github.com/scikit-image/scikit-image/blob/main/RELEASE.txt)
- [Commits](https://github.com/scikit-image/scikit-image/compare/v0.21.0...v0.22.0)

---
updated-dependencies:
- dependency-name: scikit-image
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
