+++ pyproject.toml
@@ -42,7 +42,7 @@ strict = [
   "scikit-image==0.21.0",
 ]
 
-tests = ["pytest==7.4.2", "pytest-cov==4.1.0"]
+tests = ["pytest==7.4.3", "pytest-cov==4.1.0"]
 
 
 [tool.setuptools.dynamic]
