### 2023-11-13 #5941 @ireaml
Hi @rkurchin, I've finished my review. This is a very high-quality code, with a strong motivation, easy to use and very well documented. It will be very useful for the defect community and I would be very happy to recommend publication in JOSS.

For the author, @jmmshn, these are my suggestions:

### Regarding the paper
**Line 10**: Ref 2 doesn’t display correctly. Further, it would be good to also cite the other existing packages that have been developed to pre- or post-process defect calculations, including:

    - Arrigoni, M., & Madsen, G. K. H. Spinney: Post-processing of first-principles calculations of point defects in semiconductors with Python. Computer Physics Communications, 264, 107946 (https://doi.org/10.1016/j.cpc.2021.107946)
    - Broberg, D., Medasani, B., Zimmermann, N. E. R., Yu, G., Canning, A., Haranczyk, M., Asta, M., & Hautier, G. PyCDT: A Python toolkit for modeling point defects in semiconductors and insulators. Computer Physics Communications, 226, 165–179. (https://doi.org/10.1016/j.cpc.2018.01.004)
    - Goyal, A., Gorai, P., Peng, H., Lany, S., & Stevanovi’c, V. (2017). A computational framework for automation of point defect calculations. Computational Materials Science, 130, 1–9. (https://doi.org/10.1016/j.commatsci.2016.12.040)
    - Muy, S., Johnston, C. & Marzari, N. AiiDA-defects: an automated and fully reproducible workflow for the complete characterization of defect chemistry in functional materials. Electron. Struct., 5, 024009
    - Kavanagh, S. R., Walsh, A. & Scanlon, D. O.. Defect Oriented Python Environment Distribution (DOPED), https://github.com/SMTG-Bham/doped
#### Minor typos:
**Line 16**: “most users of **a** defect analysis package**s**” -> “most users of defect analysis packages”
**Line 44**: “define **a** point defect**s**” -> “define point defects”
**Line 60**: Figure 1 appears twice in its caption (“Figure 1: Figure 1. Defect generation.”)
**Line 70**: “has **to** potential to save” -> has **the** potential to save
**Line 74**: A focus of the present package**s** is -> A focus of the present package is
**Line 75**: Would suggest rewording the sentence below to split in two. “This is important since the overwhelming majority of users of the defect analysis software will not need to run thousands of defect calculations so time invested in learning a high-throughput workflow framework only worthwhile for true high-throughput studies.”
**Line 78**: “**these** package” -> “**this** package”

### Regarding the docs
1.  For the tutorial on Shockley-Read-Hall Recombination, would be good to mention that the package approach assumes the harmonic approximation, and that this can be a bad approximation for many defects that undergo strong lattice relaxations, including:
      -  Si substitution on a Ga site in GaAs. [This study](https://journals.aps.org/prb/abstract/10.1103/PhysRevB.100.041202) describes the issues with the harmonic approximation for defects that undergo strong reconstructions - might be a good example to mention in the Shockley-Read-Hall Recombination tutorial as an example of the limitations of the harmonic approximation.
      - The [Cadmium vacancy](https://pubs.acs.org/doi/10.1021/acsenergylett.1c00380?goto=supporting-info) and [Tellurium interstitial](https://pubs.rsc.org/en/content/articlelanding/2022/fd/d2fd00043a) in CdTe
      - The [iodine interstitial](https://pubs.acs.org/doi/full/10.1021/jacs.1c03064) in CH3NH3PbI3
Accordingly, would be useful to reference the users to [CarrierCapture](https://github.com/WMD-group/CarrierCapture.jl) as a package that can describe anharmonic PES's.
2. Would suggest to include community guidelines in the `README/docs` to inform users on how to contribute to the code/report issues.
3. Think it might be useful for the users if the API sections includes links to the source code (so that users can quickly see what type the function inputs are etc)
#### Small typos:
4.	In https://materialsproject.github.io/pymatgen-analysis-defects/content/defining-defects.html#defining-defect-complexes blow -> below.
5.	In https://materialsproject.github.io/pymatgen-analysis-defects/content/nonradiative.html#evaluating-the-electron-phonon-matrix-element-with-paws :
      - "The coupling **between of** the atomic displacements" -> "The coupling **between** the atomic displacements".
      - "Note **the the** initial" -> "Note the initial"
      - "This expression above can be **calculated VASP** version 6" -> "This expression above can be calculated with VASP version 6"
6. In https://materialsproject.github.io/pymatgen-analysis-defects/content/photo-conduct.html#photo-conductivity
      -  “and **a allowed** users to control” -> “and allowed users to control”
7.	https://materialsproject.github.io/pymatgen-analysis-defects/content/formation-energy.html#getting-the-elemental-chemical-potentials
      -  "free variable**s** that indicate**s**" -> "free variables that indicate"
      -  "Here since the defect must **for** in GaN" -> "Here since the defect must **form** in GaN"
8.	In https://materialsproject.github.io/pymatgen-analysis-defects/content/formation-energy.html#finite-size-corrections, the last cell showing the plots for the planar-averaged potentials doesn't display
9.	In https://materialsproject.github.io/pymatgen-analysis-defects/content/freysoldt-correction.html#freysoldt-correction “The result**s** from sxdefectalign **is show** on the right column.” -> “The results from sxdefectalign **are shown** on the right column.”

Overall, really excellent! :)
