### 2023-12-08 bb4458d
allow linelabels, bulk_stability
