### 2023-12-08 #5941 @atimmins7
@rkurchin @jmmshn - The pymatgen-analysis-defects package is well explained by the document, the provided tutorial clearly demonstrates high-level functionality, and the code is clear and easy to follow. As a note, a number of the features rely heavily on VASP, which I do not use or have access to. As a result, I have relied on the example files in the tutorial to validate that functionality is as stated. Where possible, I used Quantum Espresso to test package functionality.

Some typos have already been highlighted and have not been repeated here. The notes below are minor, mostly about improving phrasing or code comments to improve clarity.

### Paper
Minor citation error in line 10, "Huang et al. (2022),"

### Documentation
I did not see reference to community guidelines as defined the rubric.

### Tutorial

**Quickstart**
1. Details clear, functions work as expected with provided examples

**Defining Defects**
1. Details clear, functions work as expected with provided examples
2. The caveats/notes provided are helpful to understand possible limits of the I*implementation
3. Additional tests worked as expected.

**Defect Finder**
1. Small typo in the explanation of the first code block: "The last for `q=3` demonstrates". However, the last iteration occurs at  q=2.
2. The logic is well explained, and the provided tutorial works as expected.
3. Defect Finder worked correctly with additional examples

**Shockley-Read-Hall Recombination**
1. The definition of 'Q' as a configuration coordinate is explained in the provided paper, but is not explained in the tutorial. This was initially slightly confusing and it may be helpful to clarify what it represents.
2. The code worked as expected and the notes about the process were clear.

**Photo Conductivity**
1. Small type in first paragraph: "and a allowed users" -> "and allowed users"
2. Functionality is clearly explained and tutorial works as expected.

**Formation Energy Diagrams**
1. Getting the elemental chemical potentials. It would be more accurate to specify that the numbers listed for the chemical potential vertices are the vertices of the **change** in chemical potential with respect to the bulk state.
2. The last section of "Formation Energy Diagrams" is Finite-Size Corrections. In the section,  there is an explanation of how to plot the FNV correction, but the documentation includes an error, "DefectEntry' object has no attribute 'correction_metadata". It is unclear why this error is included in the documentation and what change needs to be done to get the code to operate as expected
3.  Perhaps this latter section should briefly explain the correction and then point to the dedicated tutorial page for the FNV correction

**Freysoldt Correction**
1. It may be helpful for 'comparison_plots' to include some labeling of the charge (q = -1,0,1,2) and code (pymatgen-analysis-defects vs. sxdefectalign) for the provided plots
2. Functionality is clearly explained and tutorial works as expected

### Code Review

The code is clear and well documented. It was possible to interpret the purpose of each class/function from the code itself and the included comments. So minor fixes have been included to improve the clarity of the logic.

Some internal functions do not have explanations. This may be intentional. However in case it was a mistake, any missing explanations have been noted below.

**ccd.py**
1. Code clear and well documented.
2. Function 'from_vaspruns' (lines 143-152) - the 'spin_index' index is not listed in the explained arguments
3. Function 'from_vaspruns' (lines 143-152) - the 'band_window' index is not listed in the explained arguments
4. Function '_get_ediff': the argument 'band_structure' is listed as an argument, but isn't an argument

**core.py**
1. Function 'symmetrized_structure' (line 312) - Explanation of function appears to be for get_multiplicity function
2. Function Substitution > __init__ () - Comments do not reference kwargs
3. Function Interstitial > __init__ () - Comments do  not reference equivalent_sites or kwargs
4. Function '_set_selective_dynamics' (line 890) - No explanation/comments for function
5. Function 'center_structure' (line 968) - No explanation of arguments
6. Function '_get_el_changes_from_structures' (line 981) - explanation and actual code differ on the object returned (string vs dictionary)
7. Function '_get_defect_name' (line 1015) - Explanation of function references different function arguments

**finder.py**
1. Function 'get_site_vecs'  (line 236) - No explanation of arguments
2. Function 'cosine_similarity' (line 245) - No explanation of arguments
3. Function '_get_broundary' (line 277) - No explanation of arguments

**generators.py**
1. Function '_space_group_analyzer' (line 35) - no explanation of arguments
2. Function 'get_defects' (line 49) - no explanation of arguments
3. Function InterstitialGenerator > 'generate' (line 246-249) - No explanation of the equivalent_positions argument
4. Class ChargeInterstitialGenerator (line 397) - No explanation of max_insertions

**recombination.py**
1. Function 'fact' (line 29) - No explanation of arguments
2. Function 'herm' (line 39) - No explanation of arguments
3. Function 'get_mn' (line 171) - No explanation of dQ

**supercells.py**
1. Function '_ase_cubic' (line 125) - Does not include explanation of function or arguments

**thermo.py**
1. Class DefectEntry - No reference in the explanation to the 'entry_id' attribute
2. Function FormationEnergyDiagram > with_atomic_entries (line 283) - Explanation references band_gap/inc_inf_values, but not in function definition
3. Function FormationEnergyDiagram > get_transitions (line 517) - Explanation does not include x_min or x_max
4. Function FormationEnergyDiagram > get_formation_energy (line 545) - Explanation does not include chempot_dict
5. Function MultiFormationEnergyDiagram>with_atomic_entries (line 667) - No explanation of arguments
6. Function fermi_dirac (line 1005) - No explanation of arguments
7. Function plot_formation_energy_diagrams (line 1015) - The arguments 'legend_loc' and 'show_legend' not listed in explanation
8. Function _get_line_color_and_style (line 1220) - No explanation of arguments

**utils.py**
1. Function eV_to_k (line 121) - typo in argument explanation, should be 'energy'
2. Function converge (line 195) - No explanation of arguments
3. Function get_local_extrema (line 267) - Explanation does not include chgcar
4. Function remove_collisions (line 302) - Explanation does not include 'structure'
5. Function get_ipr_in_window (line 839) - 'k_index' not an allowed argument. Function argument 'band_window' does not mention the argument mentioned in comment, 'band_range'
6. Function get_labeled_inserted_structure (line 942) - explanation does not reference the 'sm' StructureMatcher argument
7. Function _group_docs_by_structure (line 1003) - Function has argument 'get_structure', but explanation says 'get_hash'

**Corrections > freystoldt.py**
1. Function get_freysoldt_correction (line 41) - Explanation does not mention argument 'lattice'
2. Function perform_pot_corr (line 243) - Difference between function argument and comment argument (defect_frac_coords vs defect_frac_position), explanation does not mention 'dielectric', 'q_model', or 'mad_tol'

**Plotting > optics.py**
1. Function get_bs_eigenvalues (line 115) - Explanation does not reference 'other_defect_bands' or 'shift_eig'
2. Function _plot_eigs (line 170) - No explanation of arguments
3. Function _get_dataframe (line 284) - No explanation of arguments
4. Plotting > phases.py
5. Function _convex_hull_2d (line 77) - argument 'tol' is never used
