### 2023-12-11 71b0343
Minor fixes/Dropped 3.8 (#159)
* allow linelabels, bulk_stability

* cover
