+++ pyproject.toml
@@ -6,9 +6,9 @@ requires = ["setuptools >= 42", "versioningit ~= 1.0", "wheel"]
 authors = [{ name = "Jimmy-Xuan Shen", email = "jmmshn@gmail.com" }]
 classifiers = [
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
   "Development Status :: 5 - Production/Stable",
   "Intended Audience :: Science/Research",
   "Operating System :: OS Independent",
@@ -26,7 +26,7 @@ keywords = ["high-throughput", "automated", "dft", "defects"]
 license = { text = "modified BSD" }
 name = "pymatgen-analysis-defects"
 readme = "README.rst"
-requires-python = '>=3.8'
+requires-python = '>=3.9'
 
 [project.optional-dependencies]
 finder = ["dscribe>=2.0.0"]
@@ -74,7 +74,7 @@ line-length = 88
 extend-ignore = "E203, W503, E501, F401, RST21"
 max-line-length = 120
 max-doc-length = 120
-min-python-version = "3.8.0"
+min-python-version = "3.9.0"
 rst-roles = "class, func, ref, obj"
 select = "C, E, F, W, B, B950"
 
