+++ paper/paper.md
@@ -30,7 +30,7 @@ Multiple software packages exist to automate the simulation of point defects [@B
 2. Building large, persistent databases of point defects that are extensible to new calculations over time.
 
 Since the combinatorics of point defects in crystalline materials can be extremely large, it is important to have a software package that can be easily integrated into high-throughput workflows to manage these complex calculations.
-However, most users of a defect analysis packages will not need to run thousands of calculations so it is important to have code focused purely on the defect analysis and relegate the high-throughput workflow aspect to a separate package.
+However, most users of defect analysis packages will not need to run thousands of calculations so it is important to have code focused purely on the defect analysis and relegate the high-throughput workflow aspect to a separate package.
 Additionally, a well-known problem in the simulation of point defects is the fact that current structure optimization techniques can miss the ground state structure based on the initial guess in a sizable minority of cases, so the ability to easily re-visit and re-optimize structures is crucial to building a reliable database of point defects.
 Towards that end, we have developed a Python package, `pymatgen-analysis-defects`, and integrated it with the popular `atomate2` workflow framework to provide a complete set of tools for simulating, analyzing, and managing the results of point defect calculations.
 
@@ -49,7 +49,7 @@ https://materialsproject.github.io/pymatgen-analysis-defects/intro.html
 
 ## Defect Definition
 
-A core feature of `pymatgen-analysis-defects` is the ability to define a point defects automatically.
+A core feature of `pymatgen-analysis-defects` is the ability to define point defects automatically.
 While symmetry analysis on the atomic structure alone is usually enough to define the distinct substitutional and vacancy defects, we found that the electronic charge density was the most effective at placing the interstitial defect at symmetry-inequivalent positions.
 A basic example of creating a full list of defects is shown below:
 
@@ -80,7 +80,7 @@ N intersitial site at [0.35,0.65,0.69]
 In the code above, we query the materials project database for the charge density object which contains the information about the bulk structure as well as the electronic charge density.
 Using the `generate_all_native_defects` function, we can generate a list of all of the native point defects for this structure.
 
-![Figure 1. Defect generation.](fig1.png){ width=35% }
+![Defect generation.](fig1.png){ width=35% }
 
 ## Defect Simulation Workflow
 
@@ -109,12 +109,13 @@ The code above will generate a `Flow` object that contains all of the instructio
 # Statement of need
 
 Due to the large simulation cell and the higher-cost density functionals required for defect simulations, the computational cost of defect calculations is often orders of magnitude higher than that of bulk calculations.
-As such, managing and curating the results of the defect calculations generated by a single user has to potential to save a significant amount of computational resources.
+As such, managing and curating the results of the defect calculations generated by a single user has the potential to save a significant amount of computational resources.
 Moreover, eventually building a high-quality, persistent defects database will significantly reduce the computational cost of defect calculations for the entire community.
 The present software package is designed to facilitate both of these goals.
 
-A focus of the present packages is also to provide a base library for the analysis of point defects without invoking any high-throughput workflow frameworks.
-This is important since the overwhelming majority of users of the defect analysis software will not need to run thousands of defect calculations so time invested in learning a high-throughput workflow framework only worthwhile for true high-throughput studies.
+A focus of the present package is also to provide a base library for the analysis of point defects without invoking any high-throughput workflow frameworks.
+This is important since the overwhelming majority of users of the defect analysis software will not need to run thousands of defect calculations.
+As such, time invested in learning a high-throughput workflow framework only worthwhile for true high-throughput studies.
 Even though this packages was designed with high-throughput in mind and developed alongside a high-throughput workflow framework, it is not dependent on any particular workflow framework and can be used as a standalone analysis package.
 
 # Acknowledgements
