+++ .github/workflows/testing.yml
@@ -55,7 +55,10 @@ jobs:
           pip install -e .[tests]
 
       - name: Test
-        run: pytest --cov=pymatgen --cov-report=xml
+        run: |
+          pytest --cov=pymatgen --cov-report=xml
+          pytest --nbmake ./docs/source/content
+
 
       - uses: codecov/codecov-action@v1
         if: matrix.python-version == 3.10
