+++ .github/workflows/testing.yml
@@ -54,11 +54,11 @@ jobs:
           pip install -e .[strict]
           pip install -e .[tests]
 
-      - name: Test
-        run: |
-          pytest --cov=pymatgen --cov-report=xml
-          pytest --nbmake ./docs/source/content
+      - name: Test Modules
+        run: pytest --cov=pymatgen --cov-report=xml
 
+      - name: Test Notebooks
+        run: pytest --nbmake ./docs/source/content
 
       - uses: codecov/codecov-action@v1
         if: matrix.python-version == 3.10
