+++ paper/paper.bib
@@ -909,3 +909,16 @@
 	publisher = {AIP Publishing},
 	doi = {10.1063/1.4964831}
 }
+
+@article{Arrigoni2021Jul,
+	author = {Arrigoni, Marco and Madsen, Georg K. H.},
+	title = {{Spinney: Post-processing of first-principles calculations of point defects in semiconductors with Python}},
+	journal = {Comput. Phys. Commun.},
+	volume = {264},
+	pages = {107946},
+	year = {2021},
+	month = jul,
+	issn = {0010-4655},
+	publisher = {North-Holland},
+	doi = {10.1016/j.cpc.2021.107946}
+}
