+++ paper/paper.bib
@@ -922,3 +922,16 @@
 	publisher = {North-Holland},
 	doi = {10.1016/j.cpc.2021.107946}
 }
+
+@article{Goyal2017Apr,
+	author = {Goyal, Anuj and Gorai, Prashun and Peng, Haowei and Lany, Stephan and Stevanovi{\ifmmode\acute{c}\else\'{c}\fi}, Vladan},
+	title = {{A computational framework for automation of point defect calculations}},
+	journal = {Comput. Mater. Sci.},
+	volume = {130},
+	pages = {1--9},
+	year = {2017},
+	month = apr,
+	issn = {0927-0256},
+	publisher = {Elsevier},
+	doi = {10.1016/j.commatsci.2016.12.040}
+}
