+++ paper/paper.bib
@@ -935,3 +935,13 @@
 	publisher = {Elsevier},
 	doi = {10.1016/j.commatsci.2016.12.040}
 }
+
+@misc{Smtg-Bham2023Dec,
+	author = {Mosquera-Lois, Irea and Kavanagh, Se{\ifmmode\acute{a}\else\'{a}\fi}n R. and Walsh, Aron and Scanlon, David O.},
+	title = {{doped}},
+	journal = {GitHub},
+	year = {2023},
+	month = dec,
+	note = {[Online; accessed 12. Dec. 2023]},
+	url = {https://github.com/SMTG-Bham/doped}
+}
