+++ paper/paper.bib
@@ -937,7 +937,7 @@
 }
 
 @misc{Smtg-Bham2023Dec,
-	author = {Mosquera-Lois, Irea and Kavanagh, Se{\ifmmode\acute{a}\else\'{a}\fi}n R. and Walsh, Aron and Scanlon, David O.},
+	author = {Kavanagh, Se{\ifmmode\acute{a}\else\'{a}\fi}n R. and Mosquera-Lois, Irea and Walsh, Aron and Scanlon, David O.},
 	title = {{doped}},
 	journal = {GitHub},
 	year = {2023},
