+++ README.rst
@@ -79,7 +79,15 @@ To make maintenance easier, please use a workflow similar to the automated CI
 
 Specifically, please make sure to run the following commands for linting:
 
-..  code-block:: bash
+.. code-block:: bash
+
+    $ pip install -r requirements.txt
+    $ pre-commit install
+    $ pre-commit run --all-files
+
+
+.. code-block:: bash
+
     $ pip install -e .[strict]
     $ pip install -e .[dev]
     $ pre-commit install
