+++ README.rst
@@ -96,6 +96,7 @@ Specifically, please make sure to run the following commands for linting:
 And run these commands for testing:
 
 ..  code-block:: bash
+
     $ pip install -e .[strict]
     $ pip install -e .[tests]
     $ pytest --cov=pymatgen --cov-report=xml
