+++ pymatgen/analysis/defects/ccd.py
@@ -761,7 +761,7 @@ def _get_ks_ediff(
     for ispin, eigs in bandstructure.bands.items():
         spin_index = 0 if ispin == Spin.up else 1
         res[ispin] = np.zeros_like(eigs)
-        for ikpt, kpt in enumerate(bandstructure.kpoints):
+        for ikpt, _kpt in enumerate(bandstructure.kpoints):
             iband = b_at_kpt_and_spin.get((ikpt, spin_index), None)
             if iband is None:
                 continue
