+++ .pre-commit-config.yaml
@@ -31,7 +31,7 @@ repos:
   hooks:
   - id: flake8
     entry: pflake8
-    files: ^src/
+    files: ^pymatgen/
     additional_dependencies:
     - pyproject-flake8==6.1.0
     - flake8-bugbear==22.12.6
