+++ .pre-commit-config.yaml
@@ -1,56 +1,21 @@
+exclude: ^(docs|tests)
+
+ci:
+  autoupdate_schedule: monthly
+  skip: [mypy]
+  autofix_commit_msg: pre-commit auto-fixes
+  autoupdate_commit_msg: pre-commit autoupdate
+
 repos:
-- repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v4.4.0
+- repo: https://github.com/astral-sh/ruff-pre-commit
+  # Ruff version.
+  rev: v0.1.8
   hooks:
-  - id: check-yaml
-  - id: fix-encoding-pragma
-    args:
-    - --remove
-  - id: end-of-file-fixer
-  - id: trailing-whitespace
-- repo: https://github.com/PyCQA/autoflake
-  rev: v2.2.1
-  hooks:
-  - id: autoflake
-- repo: https://github.com/psf/black
-  rev: 23.9.1
-  hooks:
-  - id: black
-- repo: https://github.com/asottile/blacken-docs
-  rev: 1.16.0
-  hooks:
-  - id: blacken-docs
-    additional_dependencies: [black]
-    exclude: README.md
-- repo: https://github.com/pycqa/isort
-  rev: 5.12.0
-  hooks:
-  - id: isort
-- repo: https://github.com/pycqa/flake8
-  rev: 6.1.0
-  hooks:
-  - id: flake8
-    entry: pflake8
-    files: ^pymatgen/
-    additional_dependencies:
-    - pyproject-flake8==6.1.0
-    - flake8-bugbear==22.12.6
-    - flake8-typing-imports==1.14.0
-    - flake8-docstrings==1.6.0
-    - flake8-rst-docstrings==0.3.0
-    - flake8-rst==0.8.0
-- repo: https://github.com/pycqa/pydocstyle
-  rev: 6.3.0
-  hooks:
-    - id: pydocstyle
-      additional_dependencies: ["tomli==2.0.1"]
-- repo: https://github.com/pre-commit/pygrep-hooks
-  rev: v1.10.0
-  hooks:
-  - id: python-use-type-annotations
-  - id: rst-backticks
-  - id: rst-directive-colons
-  - id: rst-inline-touching-normal
+    # Run the linter.
+    - id: ruff
+      args: [ --fix ]
+    # Run the formatter.
+    - id: ruff-format
 - repo: https://github.com/pre-commit/mirrors-mypy
   rev: v1.5.1
   hooks:
