+++ pymatgen/analysis/defects/corrections/freysoldt.py
@@ -9,10 +9,6 @@ import matplotlib.pyplot as plt
 import numpy as np
 import scipy
 from numpy.typing import ArrayLike
-from pymatgen.core import Lattice
-from pymatgen.io.vasp.outputs import Locpot
-from scipy import stats
-
 from pymatgen.analysis.defects.finder import DefectSiteFinder
 from pymatgen.analysis.defects.utils import (
     CorrectionResult,
@@ -23,6 +19,9 @@ from pymatgen.analysis.defects.utils import (
     generate_reciprocal_vectors_squared,
     hart_to_ev,
 )
+from pymatgen.core import Lattice
+from pymatgen.io.vasp.outputs import Locpot
+from scipy import stats
 
 __author__ = "Jimmy-Xuan Shen, Danny Broberg, Shyam Dwaraknath"
 __copyright__ = "Copyright 2022, The Materials Project"
