+++ tests/test_ccd.py
@@ -3,14 +3,13 @@ from collections import namedtuple
 import numpy as np
 import pandas as pd
 import pytest
-
 from pymatgen.analysis.defects.ccd import (
     HarmonicDefect,
     _get_wswq_slope,
     plot_pes,
 )
-from pymatgen.io.vasp.outputs import Waveder
 from pymatgen.analysis.defects.plotting.optics import plot_optical_transitions
+from pymatgen.io.vasp.outputs import Waveder
 
 
 @pytest.fixture(scope="session")
