+++ tests/test_finder.py
@@ -1,7 +1,6 @@
 def test_defect_finder(test_dir):
-    from pymatgen.core import IStructure
-
     from pymatgen.analysis.defects.finder import DefectSiteFinder
+    from pymatgen.core import IStructure
 
     base = IStructure.from_file(test_dir / "GaN.vasp")
 
