+++ .pre-commit-config.yaml
@@ -1,11 +1,5 @@
 exclude: ^(docs|tests)
 
-ci:
-  autoupdate_schedule: monthly
-  skip: [mypy]
-  autofix_commit_msg: pre-commit auto-fixes
-  autoupdate_commit_msg: pre-commit autoupdate
-
 repos:
 - repo: https://github.com/astral-sh/ruff-pre-commit
   # Ruff version.
