### 2023-12-14 564210e
Major update to Linting (#162)
* notbooks:

* linting

* update configs

* linting

* linting

* linting

* linting

* cover
