+++ pymatgen/analysis/defects/thermo.py
@@ -6,27 +6,32 @@ import logging
 from dataclasses import dataclass, field
 from itertools import chain, groupby
 from pathlib import Path
-from typing import Any, Callable, Dict, List, Optional
+from typing import TYPE_CHECKING, Any, Callable, Dict, List, Optional
 
 import numpy as np
 from matplotlib import pyplot as plt
 from monty.json import MSONable
-from numpy.typing import ArrayLike, NDArray
 from pymatgen.analysis.chempot_diagram import ChemicalPotentialDiagram
+from pymatgen.analysis.defects.core import Defect, NamedDefect
+from pymatgen.analysis.defects.corrections.freysoldt import get_freysoldt_correction
+from pymatgen.analysis.defects.finder import DefectSiteFinder
+from pymatgen.analysis.defects.utils import get_zfile, group_docs
 from pymatgen.analysis.phase_diagram import PhaseDiagram
 from pymatgen.analysis.structure_matcher import ElementComparator, StructureMatcher
-from pymatgen.core import Composition, Element, Structure
-from pymatgen.electronic_structure.dos import Dos, FermiDos
-from pymatgen.entries.computed_entries import ComputedEntry, ComputedStructureEntry
+from pymatgen.core import Composition, Element
+from pymatgen.electronic_structure.dos import FermiDos
+from pymatgen.entries.computed_entries import ComputedEntry
 from pymatgen.io.vasp import Locpot, Vasprun
 from scipy.constants import value as _cd
 from scipy.optimize import bisect
 from scipy.spatial import ConvexHull
 
-from pymatgen.analysis.defects.core import Defect, NamedDefect
-from pymatgen.analysis.defects.corrections.freysoldt import get_freysoldt_correction
-from pymatgen.analysis.defects.finder import DefectSiteFinder
-from pymatgen.analysis.defects.utils import CorrectionResult, get_zfile, group_docs
+if TYPE_CHECKING:
+    from numpy.typing import ArrayLike, NDArray
+    from pymatgen.analysis.defects.utils import CorrectionResult
+    from pymatgen.core import Structure
+    from pymatgen.electronic_structure.dos import Dos
+    from pymatgen.entries.computed_entries import ComputedStructureEntry
 
 __author__ = "Jimmy-Xuan Shen, Danny Broberg, Shyam Dwaraknath"
 __copyright__ = "Copyright 2022, The Materials Project"
@@ -441,7 +446,7 @@ class FormationEnergyDiagram(MSONable):
         Args:
             defect_entry:
                 The defect entry for which the formation energy is computed.
-            chem_pots:
+            chempots:
                 A dictionary of chemical potentials for each element.
 
         Returns:
@@ -505,7 +510,7 @@ class FormationEnergyDiagram(MSONable):
         """Get the lines for the formation energy diagram.
 
         Args:
-            chempot_dict:
+            chempots:
                 A dictionary of the chemical potentials (referenced to the elements)
                 representations a vertex of the stability region of the chemical
                 potential diagram.
@@ -524,7 +529,7 @@ class FormationEnergyDiagram(MSONable):
         return lines
 
     def get_transitions(
-        self, chempots: dict, x_min: float = 0, x_max: float = 10
+        self, chempots: dict, x_min: float = 0, x_max: float | None = None
     ) -> list[tuple[float, float]]:
         """Get the transition levels for the formation energy diagram.
 
@@ -533,10 +538,14 @@ class FormationEnergyDiagram(MSONable):
         point respectively.
 
         Args:
-            chempot_dict:
+            chempots:
                 A dictionary of the chemical potentials (referenced to the elements)
                 representations a vertex of the stability region of the chemical
                 potential diagram.
+            x_min:
[…]
