### 2023-12-14 a28b25a
More linting (#163)
* notbooks:

* linting

* update configs

* linting

* linting

* linting

* linting

* cover

* lint

* lint

* lint

* lint

* lint

* lint

* lint

* lint

* lint

* lint

* lint
