### 2023-12-14 #5941 @jmmshn
Hi  @atimmins7 @ireaml and @rkurchin,
Thanks again for the feedback!
Some of my linting and notebook testing were not optimally configured.
I have fixed the issues identified above and put things in place to catch similar problems in the future.
The repo is ready to be examined again!
Thanks!
