### 2023-12-21 #5941 @atimmins7
Well noted! Thank you for letting me know.

@jmmshn @rkurchin - I have reviewed the changes and only noticed a few small aesthetic typos. As a result, I support final publication and have updated my checklist accordingly. Although these are not blocking items, the following typos might be fixed to improve clarity.

1. In the paper's first paragraph, the citation ", (Huang et al. 2022,)" leaves the comma after 2022 inside the parentheses and "(Kavanagh et al. 2023 )however" includes the space after 2023 inside the parentheses.
2. In the 'Defect Finder' tutorial, there is a reference to charge state q = 3: "The last for `q=3` demonstrates". However, this line appears to be referencing the last iteration of the loop above, which occurs at  q=2.
