+++ pyproject.toml
@@ -42,7 +42,7 @@ strict = [
   "scikit-image==0.21.0",
 ]
 
-tests = ["pytest==7.4.2", "pytest-cov==4.1.0", "nbmake==1.4.6"]
+tests = ["pytest==7.4.4", "pytest-cov==4.1.0", "nbmake==1.4.6"]
 
 
 [tool.setuptools.dynamic]
