### 2024-01-11 eaacfa2
Patch Numba Dep (#167)
* allow linelabels, bulk_stability

* cover

* docs

* njit

* njit
