+++ pyproject.toml
@@ -34,7 +34,7 @@ dev = ["pre-commit>=2.12.1"]
 docs = [
   "jupyter-book>=0.13.1",
 ]
-pydefect = ["pydefect>=0.6.2"]
+optional = ["pydefect>=0.6.2", "dscribe>=2.0.0", "numba"]
 
 strict = [
   "pymatgen==2023.10.11",
