### 2024-01-11 #5941 @rkurchin
Some editorial notes:
* first paragraph: there are spaces missing before the citation on line 9 and after the last citation on line 11
* In the numbered lists on lines 13-15 and 35-41, either omit punctuation at the end of each number, or change first one to a comma and add "and" afterwards
* Also, add an "of" after "Integration" on line 13
* line 16: combinatorics being "large" reads oddly to me – I would either change the adjective to something like "daunting" or add some verbiage like "lead to large number of calculations"
* line 19: add comma after "calculations"
* 33: "quantum recombination" feels like a weird phrase (is there a "classical recombination" analogue?) – unless you feel it means something specific, I would suggest changing it to "carrier recombination"
* 59: I believe "materials project" should be capitalized? Also, add a comma after "object"
* 60: add a comma after "structure"
* 66: rephrase to something like "can be sent to an HPC job manager" or "can be sent to the job manager on an HPC system"

I also think a reorganization could help the flow of the paper. I would argue that the start of your Summary section (up through line 15) is really also part of the statement of need, while the second paragraph of your current Statement of Need is somewhat redundant with things you've said earlier. I would say you could trim down or even remove that second paragraph and move the Statement of Need (along with the paragraph on lines 69-75) to the top, wrapping it into lines 6-15. Then have the Summary section come right afterwards.

Let me know if any of this doesn't make sense!
