### 2024-01-12 #5941 @jmmshn
> I would say you could trim down or even remove that second paragraph and move the Statement of Need (along with the paragraph on lines 69-75) to the top,

@rkurchin, I wasn't aware that the *Statement of Need* was allowed to go on top.  I agree that will help with the flow a lot.
