+++ paper/paper.md
@@ -20,7 +20,7 @@ date: August 2023
 bibliography: paper.bib
 ---
 
-# Summary
+# Statement of need
 
 Point defects can often determine the properties of semiconductor and optoelectronic materials.
 Simulation of point defects is one of the most complex workflows in computational materials science, involving extensive pre- and post-processing of the structural and electronic structure data [@CGWalle_defects_RMP].
@@ -28,6 +28,13 @@ Multiple software packages exist to automate the simulation of point defects inc
 
 1. Integration of but not insistence on standardized high-throughput workflow frameworks
 2. Building large, persistent databases of point defects that are extensible to new calculations over time
+
+Due to the large simulation cell and the higher-cost density functionals required for defect simulations, the computational cost of defect calculations is often orders of magnitude higher than that of bulk calculations.
+As such, managing and curating the results of the defect calculations generated by a single user has the potential to save a significant amount of computational resources.
+Moreover, eventually building a high-quality, persistent defects database will significantly reduce the computational cost of defect calculations for the entire community.
+The present software package is designed to facilitate both of these goals.
+
+# Summary
 
 Since the combinatorics of point defects in crystalline materials can be daunting, it is important to have a software package that can be easily integrated into high-throughput workflows to manage these complex calculations.
 However, most users of defect analysis packages will not need to run thousands of calculations, so it is important to have code focused purely on the defect analysis and relegate the high-throughput workflow aspect to a separate package.
@@ -106,17 +113,6 @@ flow = Flow(jobs)
 The code above will generate a `Flow` object that contains all of the instructions to dynamically create all of the required defect calculations, which can be sent to the job manager on an HPC system.
 
 
-# Statement of need
-
-Due to the large simulation cell and the higher-cost density functionals required for defect simulations, the computational cost of defect calculations is often orders of magnitude higher than that of bulk calculations.
-As such, managing and curating the results of the defect calculations generated by a single user has the potential to save a significant amount of computational resources.
-Moreover, eventually building a high-quality, persistent defects database will significantly reduce the computational cost of defect calculations for the entire community.
-The present software package is designed to facilitate both of these goals.
-
-A focus of the present package is also to provide a base library for the analysis of point defects without invoking any high-throughput workflow frameworks.
-This is important since the overwhelming majority of users of the defect analysis software will not need to run thousands of defect calculations.
-As such, time invested in learning a high-throughput workflow framework is only worthwhile for true high-throughput studies.
-Even though this package was designed with high-throughput in mind and developed alongside a high-throughput workflow framework, it is not dependent on any particular workflow framework and can be used as a standalone analysis package.
 
 # Acknowledgements
 
