+++ paper/paper.md
@@ -23,21 +23,23 @@ bibliography: paper.bib
 # Statement of need
 
 Point defects can often determine the properties of semiconductor and optoelectronic materials.
+Due to the large simulation cell and the higher-cost density functionals required for defect simulations, the computational cost of defect calculations is often orders of magnitude higher than that of bulk calculations.
+As such, managing and curating the results of the defect calculations generated by a single user has the potential to save a significant amount of computational resources.
+Moreover, eventually building a high-quality, persistent defects database will significantly reduce the computational cost of defect calculations for the entire community.
+
 Simulation of point defects is one of the most complex workflows in computational materials science, involving extensive pre- and post-processing of the structural and electronic structure data [@CGWalle_defects_RMP].
 Multiple software packages exist to automate the simulation of point defects including work from [@Broberg2018], [@Kumagai2021], [@Huang2022], [@Arrigoni2021Jul], [@Goyal2017Apr], and [@Smtg-Bham2023Dec] however, there is a lack of a code that focuses on:
 
 1. Integration of but not insistence on standardized high-throughput workflow frameworks
 2. Building large, persistent databases of point defects that are extensible to new calculations over time
 
-Due to the large simulation cell and the higher-cost density functionals required for defect simulations, the computational cost of defect calculations is often orders of magnitude higher than that of bulk calculations.
-As such, managing and curating the results of the defect calculations generated by a single user has the potential to save a significant amount of computational resources.
-Moreover, eventually building a high-quality, persistent defects database will significantly reduce the computational cost of defect calculations for the entire community.
-The present software package is designed to facilitate both of these goals.
-
 # Summary
 
 Since the combinatorics of point defects in crystalline materials can be daunting, it is important to have a software package that can be easily integrated into high-throughput workflows to manage these complex calculations.
 However, most users of defect analysis packages will not need to run thousands of calculations, so it is important to have code focused purely on the defect analysis and relegate the high-throughput workflow aspect to a separate package.
+A focus of the present package is also to provide a base library for the analysis of point defects without invoking any high-throughput workflow frameworks.
+Even though this package was designed with high-throughput in mind and developed alongside a high-throughput workflow framework, it is not dependent on any particular workflow framework and can be used as a standalone analysis package.
+
 Additionally, a well-known problem in the simulation of point defects is the fact that current structure optimization techniques can miss the ground state structure based on the initial guess in a sizable minority of cases, so the ability to easily re-visit and re-optimize structures is crucial to building a reliable database of point defects.
 Towards that end, we have developed a Python package, `pymatgen-analysis-defects`, and integrated it with the popular `atomate2` workflow framework to provide a complete set of tools for simulating, analyzing, and managing the results of point defect calculations.
 
