+++ paper/paper.md
@@ -12,6 +12,7 @@ authors:
     orcid: 0000-0002-2743-7531
     affiliation: "1" # (Multiple affiliations must be quoted)
   - name: Joel B. Varley
+    orcid: 0000-0002-5384-5248
     affiliation: "1"
 affiliations:
   - name: Lawrence Livermore National Laboratory, Livermore, California 94550, United States
