### 2024-01-19 #5941 @kyleniemeyer
Hi @jmmshn, I made some minor edits to the paper in https://github.com/materialsproject/pymatgen-analysis-defects/pull/169, can you review and merge that?

Also, please add the DOI 10.1103/RevModPhys.86.253 to reference Freysoldt et al.
