+++ paper/paper.bib
@@ -667,6 +667,7 @@
    title = {First-principles calculations for point defects in solids},
    volume = {86},
    year = {2014},
+   doi = {10.1103/RevModPhys.86.253},
 }
 
 @article{pymatgen,
